Add method to sanitize instructions
-Sanitize_instructions takes in the instruction string and splits it
based on white space
 - It then converts any numbers and into integers.
- It returns an new array of integers and instruction strings

Fix assert statement in tests, add method for +

- changed assert —> assert_equals where appropriate in test
- add method removes last to numbers from stack, sums them and pushes
the sum back to the stack

Add subract method

- subtract method removes last two integers from stack and subtracts
the smaller value from the larger then pushes the difference back to
the top of the stack.

- refactored to have the stack values be integers. They are converted
to strings when the solution is returned.

Add duplicate method

- duplicate pushes a duplicate of the last element in the stack to the
top of the stack
- added complex instructions to test —> Stack Calculator works!